Consistent discussion summaries Add a reusable discussionSummary component and use it for all discussion listings.
