+++ src/boards/boards/chat.njk
@@ -21,19 +21,13 @@ eleventyComputed:
     <ul class="discussion-list">
     {% for discussion in board.featured | pinned(board.zooniverse_id) %}
       <li>
-        <p class="pinned-discussion">
-          <a href="./discussions/{{ discussion.zooniverse_id }}">{{ discussion.title }}</a>
-        </p>
-        <p>{{ discussion.comments }} posts / {{ discussion.users }} participants</p>
-        <p>Last post {{ discussion.last_comment.created_at | date }} by {{ discussion.last_comment.user_name }}</p>
+         {% discussionSummary discussion=discussion, className='pinned-discussion' %}
       </li>
     {% endfor %}
 
     {%- for discussionID in board.links.discussions | notFeatured(board.featured) -%}
       <li>
-        <p><a href="./discussions/{{ discussions[discussionID].zooniverse_id }}">{{ discussions[discussionID].title }}</a></p>
-        <p>{{ discussions[discussionID].comments.length }} posts / {{ discussions[discussionID].users }} participants</p>
-        <p>Last post {{ discussions[discussionID].last_comment.created_at | date }} by {{ discussions[discussionID].last_comment.user_name }}</p>
+        {% discussionSummary discussion=discussions[discussionID] %}
       </li>
     {%- endfor -%}
     </ul>
